bloqueando exclusao de especialidade agendada

+++ src/main/java/com/jenry/curso/security/web/controller/MedicoController.java
@@ -62,8 +62,14 @@ public class MedicoController {
     @GetMapping({"/id/{idMed}/excluir/especializacao/{idEsp}"})
     public String excluirEspecialidadePorMedico(@PathVariable("idMed") Long idMed,
                           @PathVariable("idEsp") Long idEsp, RedirectAttributes attr){
-        service.excluirEspecialidadePorMedico(idMed,idEsp);
-        attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+
+        if (service.existeEspecialidadeAgendada(idMed, idEsp)){
+            attr.addFlashAttribute(("falha"), "Existem consultas agendadas com esta especialidade, exlusão negada.");
+        } else {
+            service.excluirEspecialidadePorMedico(idMed,idEsp);
+            attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+        }
+
         return "redirect:/medicos/dados";
     }
